+++ app/javascript/vue/components/radials/navigation/radial.vue
@@ -71,12 +71,17 @@ import RecentComponent from './components/recent.vue'
 import DestroyConfirmation from './components/DestroyConfirmation'
 import all_tasksComponent from './components/allTasks.vue'
 
-const defaultOptions = {
+const DEFAULT_OPTIONS = {
   New: 'New',
   Edit: 'Edit',
   Destroy: 'Destroy',
   Recent: 'Recent',
   Show: 'Show'
+}
+
+const CUSTOM_OPTIONS = {
+  AllTasks: 'allTasks',
+  CircleButton: 'circleButton'
 }
 
 export default {
@@ -164,7 +169,7 @@ export default {
       }
 
       if (this.metadata?.recent_url) {
-        taskSlices.push(this.addSlice(defaultOptions.Recent,
+        taskSlices.push(this.addSlice(DEFAULT_OPTIONS.Recent,
           this.recentTotal
             ? {
                 slices: [{
@@ -206,7 +211,7 @@ export default {
       const filterOptions = this.filterOptions
 
       if (!this.metadata.destroy) {
-        filterOptions.push(this.addSlice(defaultOptions.Destroy))
+        filterOptions.push(this.addSlice(DEFAULT_OPTIONS.Destroy))
       }
 
       return this.defaultSlicesTypes.filter(type => !filterOptions.includes(type)).map(type => this.addSlice(type, { link: this.defaultLinks()[type] }))
@@ -217,12 +222,12 @@ export default {
     },
 
     isPinned () {
-      return this.metadata['pinboard_item']
+      return this.metadata?.pinboard_item
     },
 
     middleButton () {
       return {
-        name: 'circleButton',
+        name: CUSTOM_OPTIONS.CircleButton,
         radius: 30,
         icon: {
           url: Icons.Pin,
@@ -243,16 +248,15 @@ export default {
       display: false,
       globalIdSaved: undefined,
       metadata: undefined,
-      title: 'Radial object',
+      title: 'Radial navigation',
       deleted: false,
       showDestroyModal: false,
       recentTotal: 0,
-      customOptions: ['alltasks', 'circleButton'],
       defaultSlicesTypes: [
-        defaultOptions.New,
-        defaultOptions.Destroy,
-        defaultOptions.Edit,
-        defaultOptions.Show
+        DEFAULT_OPTIONS.New,
+        DEFAULT_OPTIONS.Destroy,
+        DEFAULT_OPTIONS.Edit,
+        DEFAULT_OPTIONS.Show
       ]
     }
   },
@@ -273,39 +277,29 @@ export default {
     },
 
     selectedRadialOption ({ name }) {
-      if (Object.keys(defaultOptions).includes(name) || this.customOptions.includes(name)) {
-        this.currentView = undefined
-        switch (name) {
-          case 'circleButton':
-            this.isPinned ? this.destroyPin() : this.createPin()
-            break
-          case defaultOptions.Recent:
-            this.currentView = 'Recent'
-            break
-          case defaultOptions.Edit:
-            window.open(this.metadata['edit'] ? this.metadata.edit : `${this.metadata.resource_path}/edit`)
-            break
-          case defaultOptions.New:
-            window.open(this.metadata['new'] ? this.metadata.new : `${this.metadata.resource_path.substring(0, this.metadata.resource_path.lastIndexOf('/'))}/new`)
-            break
-          case defaultOptions.Show:
-            window.open(this.metadata.resource_path)
-            break
-          case defaultOptions.Destroy:
-            this.showDestroyModal = true
-            break
-          case 'alltasks':
-            this.currentView = 'all_tasks'
-            break
-        }
+      switch (name) {
+        case CUSTOM_OPTIONS.CircleButton:
+          this.isPinned
+            ? this.destroyPin()
+            : this.createPin()
+          break
+        case DEFAULT_OPTIONS.Recent:
+          this.currentView = 'Recent'
+          break
+        case DEFAULT_OPTIONS.Destroy:
+          this.showDestroyModal = true
+          break
+        case 'alltasks':
+          this.currentView = 'all_tasks'
+          break
       }
     },
 
     defaultLinks () {
       return {
-        [defaultOptions.Edit]: this.metadata?.edit || `${this.metadata.resource_path}/edit`,
-        [defaultOptions.New]: this.metadata?.new || `${this.metadata.resource_path.substring(0, this.metadata.resource_path.lastIndexOf('/'))}/new`,
-        [defaultOptions.Show]: this.metadata.resource_path
+        [DEFAULT_OPTIONS.Edit]: this.metadata?.edit || `${this.metadata.resource_path}/edit`,
+        [DEFAULT_OPTIONS.New]: this.metadata?.new || `${this.metadata.resource_path.substring(0, this.metadata.resource_path.lastIndexOf('/'))}/new`,
+        [DEFAULT_OPTIONS.Show]: this.metadata.resource_path
       }
     },